Use a playwright provided container for CI
By using the provided container for the current version of playwright,
we don't need to install all the browsers and system dependencies as
part of the job. We can leave it to github to cache the image, instead
of managing the caching ourselves.

The downside is the image doesn't contain some tools we were previously
relying on like pipx and make, and and the user is not able to
install packages with apt.