Add a model for participant reported mammograms
These are used to inform image readers that additional images exist,
so they can make sure that they are loaded onto the PACS (this process
is not part of the service)

This information may be incomplete or misremembered, so free text fields
allow the mammographer to capture information that will help to track
down the images.  There are also cases where the participant may not
wish to disclose where the mammogram was done, for example if it was
done in prison. In this case it is still clinically relevant to know the
mammogram was done.

If the mammogram was taken as part of the national breast screening programme
then this information may be redundant (because the service should already
has access to data from past screenings), but the data in this table is
covers all mammograms, and is also user reported, so we want
to store it separately from confirmed information.