Remove option to select a unit from the list
The idea behind this select input was to save mammographers time typing.
However, ideally it would be an autocomplete rather than a plain select,
and allow searching by address or previous names of the unit.

However, since we don't have this data available in the system at the
moment, we've decided to split this out into another ticket and postpone
it.

Without this field, mammographers will need to select "Somewhere in the
UK" and type the name.