Show participant recorded mammograms on appointment
Change LastKnownMammogramPresenter to use ParticipantRecordedMammogram
(the information provided by the participant).

Later, this will take into account recorded information from past
screenings, but we are not ready to implement this properly yet.

The previous implementation was incomplete because a screening episode
was used as a proxy for mammograms actually being done. However, the
existence of a screening episode does not guarantee the appointment
proceeded to mammograms.