Add workaround for bug in attributes macro
The attributes macro does not seem to be marking values as safe
correctly when autoescape is on. This is causing rendering issues
with the date-input component when passed a `role` attribute.

Django sets autoescape by default for security reasons, so we want to
keep this switched on for html templates.

Until we've got a proper fix, disable it specifically for this macro.