fix: remove usage of assertTemplateUsed
This assertion was doing nothing, and it's not actually possible
to use assertTemplateUsed correctly when the template engine is set to
Jinja2.

See https://code.djangoproject.com/ticket/24622
for more details.

While it would be nice to unit test this behaviour, it's unlikely to
catch anything that won't be caught by our playwright tests.