feat: implement first form for special appointments
This form processes the first step in the edit journey, and
serialises/deserialises to JSON which we can store on the participant
record for now.

The step is followed by a second step if the user selected "are any of
these reasons temporary" - this will have its own form.

I considered replacing the JSON field with a proper model, but at this
stage we're still iterating this functionality, so we decided to keep
it as flexible JSON.