feat: change json structure for extra needs
extra_needs is now a dictionary where each reason maps
to a dictionary. I'm keeping it modelled as a json field for now
as this is likely to change again and we want to be able
to iterate quickly.

The design of this banner will be updated in a future commit to match
the latest prototype, with a summary list component.