Add custom template for SplitDateField
This is a first step towards building a set of Field subclasses that are
easy to render in the template using
`{{ form.field_name.as_field_group() }}`

My hope is to replicate this pattern with CharField, ChoiceField etc. so
that we have a set of Field subclasses that do not require awkward code
in the view templates to extract the params needed for the NHS.UK frontend
components - we can just implement it once per field type.

How it works
------------

In Django, calling `form[field_name]` gives you a `BoundField` object.

In the template, you can call {{ form.field_name }} or {{ form.field_name.as_field_group }} which both
call methods on this object. Both of these calls internally render
templates. The former renders a template associated with the field's
widget, e.g. TextInput, whereas the latter renders the
`django/forms/fields.html` template, which includes form errors, labels
etc, in addition to the widget template.

The jinja version of this template looks like this

```jinja
{% if field.use_fieldset %}
  <fieldset{% if field.aria_describedby %} aria-describedby="{{ field.aria_describedby }}"{% endif %}>
  {% if field.label %}{{ field.legend_tag() }}{% endif %}
{% else %}
  {% if field.label %}{{ field.label_tag() }}{% endif %}
{% endif %}
{% if field.help_text %}<div class="helptext"{% if field.auto_id %} id="{{ field.auto_id }}_helptext"{% endif %}>{{ field.help_text|safe }}</div>{% endif %}
{{ field.errors }}
{{ field }}
{% if field.use_fieldset %}</fieldset>{% endif %}
```

It is possible to swap out either of these templates for custom ones.
However, customising at the widget level is no use to us, because in
order to call our NHS.UK frontend components, we need params for errors,
label etc. and the widget is too granular for this.

On the other hand, if we override the field template, we can do whatever
we want in the template.