Implement ConditionalField (failed experiment)
This is a MultiValueField with one ChoiceField and a number of other
fields which are conditionally revealed based on the value of the
ChoiceField.

This is intended to provide the data required to render a radios or
checkboxes component with `conditional` params.

Without this, we need to define the fields separately and implement
extra logic in the form's clean() method, and we need extra logic
in the template to construct the params for the component.

Problems
--------

How do we render this thing in the template?

The problem is that the BoundField of a MultiValueField doesn't
decompose into BoundFields for the component fields, it only
knows about their widgets.

I think this messes up error handling because we only know about errors
for the group as a whole, but when passing `conditional` to the
component we want to have errors for the parent component AND errors for
the conditional component(s).

We also need a way of accessing the right subfields in the first place,
based on the choice field value, so that we can construct the `items` param.
However, this is not possible through the BoundField API. Even if we
swapped out `bound_field_class` it would be very messy and we would
be bypassing the DictMultiWidget.