add CharField subclass that renders the `input` component
This overrides the default "field" template that looks like this (jinja
version)

     {% if field.use_fieldset %}
       <fieldset{% if field.aria_describedby %} aria-describedby="{{ field.aria_describedby }}"{% endif %}>
       {% if field.label %}{{ field.legend_tag() }}{% endif %}
     {% else %}
       {% if field.label %}{{ field.label_tag() }}{% endif %}
     {% endif %}
     {% if field.help_text %}<div class="helptext"{% if field.auto_id %} id="{{ field.auto_id }}_helptext"{% endif %}>{{ field.help_text|safe }}</div>{% endif %}
     {{ field.errors }}
     {{ field }}
     {% if field.use_fieldset %}</fieldset>{% endif %}

The `{{ field }}` in this template here triggers another template render
via the widget class. This is what allows you to swap out the widget at
form declaration time and get a different html element rendered.

Since the widget is more granular than our design system components,
we do not want to have this two-step rendering, and instead
of customising at a widget level, we will override the
field template per field type.

The idea is to have a subclass for every field type, and for that
to render the matching design system component.