only send Appointments to Notify with Booked status
The MESH Inbox will be updated when certain
actions have been performed on an Appointment -
Booked, Cancelled or Updated. The appointment
`status` field will be one of:

'B' = Booked
'C' = Cancelled
'A' = Attended
'D' = DNA

We (currently) only want to send invites to
participants if the Appointment is Booked, as that
is the scope of our private beta work.