Make `nbss_id` a unique field on Appointments in notifications
We need the ability to update statuses of
Appointments based on events coming in from NBSS -
e.g. if an Appointment is cancelled. The
Appointments coming from NBSS into the MESH inbox
will have an ID from NBSS to identify them, and we
will use that to match with Appointments already
created in our db, so it should be a unique id.