Update Appointment if it exists already
This updates an Appointment if one already exists
with the Appointment ID from NBSS

This is assuming that the NBSS ID is consistent.
It's also updating any field that isn't the NBSS
ID - so if the NHS number for instance has
changed then we will reflect that.

Will next be adding datetime stamps to show when
the model was last updated, based on the data we
have from NBSS.