Add seed_demo_data and data/demo_data.yml
* Add basic seed_demo_data command
* Add data/demo_data.yml
* Add setting to the ClinicFactory django_get_or_create args

We want to be able to populate rich, consistent example data in
non-production environments for assurance, testing and training.

This commit adds a basic mechanism for that with support for Provider,
Setting and Clinic.

Relative clinic dates has been implemented in the command for now. It
gets quite messy adding support for that inside the factory without
messing up the default starts_at and ends_at dates and the ability to
pass those in as arguments to the factory. We can look at that later if
needed.