Add Appointment, ScreeningEpisode and Participant
* Add support for creating Participants with a ScreeningEpisode and
  Appointment
* Update ParticipantFactory to get or create based on NHS number
* Update ParticipantFactory to use a sequence for NHS number

The data file has nested appointment within the clinic slot. We can use
the get or create functionality in the factories to add duplicate users
(in this case currently) and screening episodes (in the future) without
actually creating them in the database. They will just be associated
with their parent. This means that we can have episodes with more than
one appointment or participants with more than one appointment etc by
just duplicating them in the data file. That does mean 'duplication' but
it makes it easier to predict the shape of the data from reading the
datafile and that file can be managed to a good extent with an LLM so
duplication is IMO acceptable there.