support additional attributes on inputs and textareas
The design system provides params.attributes as an escape hatch for
adding additional HTML attributes that don't have their own parameters.

Django also provides an attrs argument to the widget that does
a similar thing.

We can connect the two providing that we drop the default maxlength
attribute which we don't want.