add a command to retry sending failed MessageBatches
If a MessageBatch is marked as failed, here we
have a command to retry sending it to NHS Notify.

This could be called a by a scheduled/queued job
in the near future, based on what the reason for
the failure was (e.g. if it was a temporary server
error that we can retry)