Create a queue for failed message batches
If a MessageBatch fails to send for whatever
reason, then we send it to a queue of failed
batches.

In future a worker will be able to process the
queue and determine what next steps are for the
batch - e.g. if one Message has a validation error
then we remove that Message. Or if the error was
recoverable (e.g. a server timeout) then we just
go ahead and retry.