add failed batches to a retry queue
we add any failed batches to a queue for failed
messages, and a future worker can process the
queue and determine how to handle the failure.

What we add to the queue could change - for
example error message of response to grab the
index of a failing message. (or this could be
persisted on the MessageBatch model)