Remove multi-ecosystem groups
This is still in beta, and the configuration we tried didn't work as
intended.

I was expecting to route only updates for nhsuk-frontend and
nhsuk-frontend-jinja to the "templates" multi ecosystem group, but the
PRs reference multiple updates.

These two dependencies need to be updated together, and regular
dependabot groups don't work well, because they consider each
ecosystem independently.

Since this is a low priority issue affecting only two dependencies,
let's remove this until multi-ecosystem updates is stable, and continue
to manage NHS.UK frontend updates manually for now.