Add "demo" app with persona login
This is a login screen that requires no passwords and shows
a user for each persona.

Currently, there are three personas, a clinical user who takes the
the mammogram, an administrive user who manages clinics, and
superusers who can do everything.

This will be enabled in our dev environment and review apps,
but not in production.

In future commits, we'll:

- require the user to be logged in and redirect to the persona page
  if logged out in an environment with personas enabled
- reflect the current user in the header and provide a logout button
- configure the groups with the right permissions, so that the users
  only see what they need to do their role