Update status and add errors to failed MessageBatches
When a MessageBatch fails, we look at the status
code from the response and decide whether to mark
as recoverable or unrecoverable.
We add the notify errors json to the Batch as well.

In the future, we can add retry or alert/reporting
services based on the type of failure.