Update pipelines for review apps
Github action workflows and Azure devops pipelines
The review app builds when the PR is labelled with "deploy"
The workflows now use a Github environment for each separate environment
configuration, i.e. 'review' and 'dev' as opposed to 'azure'. Each one
contains the secrets required for each ADO pipeline.
The ADO pipeline now runs for only one environment as opposed to a list
of environments. And it accepts a pull request number for review apps.