add a way to automatically assign permissions to the roles
This is done via a management command which can be run as part of our
automated deployments.

we can't do this in a data migration because permissions themselves are
generated post-migration via the post_migrate signal
(see https://docs.djangoproject.com/en/5.2/topics/auth/default/#default-permissions)

These will change whenever we add, remove, or rename models.

I was hoping we could trigger this code from the post_migrate signal so
that it would be run automatically, but I couldn't get it to work.