merge demo app into auth
This was a code review suggestion to have all the auth-related stuff in
the auth app.

I've kept a separate module for demo_views to make clear the intent
that this should not be used in production, and the view is only
added to the URLConf if PERSONAS_ENABLED is set.