fix configuration for FORM_RENDERER = "django.forms.renderers.TemplatesSetting"
As per
https://docs.djangoproject.com/en/5.2/ref/forms/renderers/#django.forms.renderers.TemplatesSetting
if we use the TemplatesSetting form renderer we need to either include
the built in template dirs in one of our template engine configurations,
OR include 'django.forms' in installed apps.

The first method didn't seem to be working as it wasn't finding
`django/forms/widgets/date.html` when rendering the participant
modeladmin. However the second method appears to work fine.