fix SHELLFLAGS typo that prevents exiting on first failure
If .ONESHELL is set, the -e option must be passed to the shell. However,
there was a typo in .SHELLFLAGS that prevented it from being picked up.

-e in bash means "Exit immediately if a command exits with a non-zero
status."

Adding to the confusion, version 3.81 of make applied the -e behaviour
regardless of the typo, but 3.82 onwards didn't, leading to different
results on different machines.

Since we have phony tasks with multiple test commands in, this was
allowing build failures to pass undetected.

I've tested the fix with a stripped down example:

    .PHONY: debug

    debug:
    	false
    	echo "This should not print"

    .ONESHELL:
    .SHELL := /bin/bash
    .SHELLFLAGS := -cex

This typo must have crept in when I stripped down the repository
template's make scripts (to get it to work), as it's not present in the
template.