add a mechanism to login in system tests
This includes steps to log in as any of the roles.

Before logging in, we need to create the test user, and as part of
that we also need to create the group (representing the role).

Unintuitively, the group may or may not exist, depending on the test
order! This is because the group *should* be created as part of
the data migrations, but `StaticServerTestCase` (which inherits from
`TransactionTestCase`) truncates the table between tests.

According to the docs
(https://docs.djangoproject.com/en/5.2/topics/testing/overview/#rollback-emulation)
the normal way to deal with this is set `serialized_rollback=True` in
the body of the test case; however, this was not working when I tried
it out.

The system tests are now split into subdirectories oriented around
the roles, with "general" for scenarios where the role doesn't
matter.