check "holding clinic" appointments are not saved
In the MESH data we receive the field "Holding
Clinic" denotes a workflow step in NBSS.

If the dat file value for this field is "Y" we
should not save the appointment data.

We don't anticipate that Birmingham's data will
have this value as they don't use this workflow
but it is known to be used by other BSOs.