remove role-based tests temporarily
We don't want these tests to be coupled to role, because the roles
are likely to change, and within the views access control is based
around permissions, not roles.

We are going to think about how we can test this better and come back to
it. Ideally we want a way of stubbing out the `user.has_perm` result,
and exercising both true/false. We also want a way to ensure that
we don't have views that do not check any permissions (which
is the case at the moment, but not the desired end state)