Process validation errors
If there is a validation error on one or more messages in
a Batch, then Notify will send the index of that Message
and information about why it is invalid.
This first pass at handling the validation errors does two things:
* Gets the Message(s) by its index (assuming the index is the same
as its index on a Batch) and removes it from its Batch
* Resends the altered Batch to the retry queue to see if it can be sent
again.