use LASTNAME, firstname (Role) format in header
This format is a convention used across clinical applications.
It used to be mandated as part of the common interface standard.
Although it is not mandated any more, we don't want to diverge from the
convention without a good reason.

Constructing `role_and_text` in the template itself is a bit awkward.
We can't put it in jinja_env.py because we need access to `request.user`.
We could create a context processor for this, but the
django docs discourage using context processors with jinja:
https://docs.djangoproject.com/en/5.2/topics/templates/#django.template.backends.jinja2.Jinja2