Fix bug in CIS2 back channel system spec
Our CIS2 OAuth client is cached in the registry of a running Django app
from the first time the get_cis2_client function is called. This
includes caching of the payload received from CIS2's JWKS endpoint.

In our system test, this means we have to ensure that there is
consistency between:
- the key pairs we generate
- the mocked response of the JWKS URI
- the cached value of the JWKS endpoint in the OAuth client

This isn't currently happening. As a result, our 'expired token'
scenario is passing but for the wrong reasons. It fails on a bad
signature error rather than an expiry error.

To correct this we:

- Move all key pair set up and request mocking into a pytest fixture
  at the top of the system test (rather than inline) so it's clearer
  what's being mocked
- Make use of yield in this fixture so we can easily specify some
  clean up
- As part of this clean up, remove the 'cis2' OAuth client from the
  registry

This results in a new client being set up for each of these test
scenarios.