rename ParticipantReportedSymptom to just Symptom
It's not important that the participant reported the symptom, just that
it was captured during/in preparation for the appointment.

It's possible the mammographer could also notice symptoms such as lumps
during an appointment that the participant hasn't reported. We may later
support mammographers recording that, in which case we would use the
same model, but perhaps add an extra field to specify the provenance.

Features captured as part of image reading will be modelled separately
later.