extend SplitDateField to support Month/Year version
This replaces the custom widget class and template with two variants
each:
one for the 3 field version and one for the 2 field version.

The SplitDateField decides which to use based on the `include_day`
argument.

I've removed the hidden version of the widget to avoid having extra code
we're not making use of anywhere.  We can reimplment this if we have a
need for it; the default behaviour is to render the `HiddenInput`
widget if hidden=True.