Add BASE_URL env var to config
On deployed environments, we need to be able to access the base url of
our app in order to construct a redirect_uri to pass on to the CIS2 auth
service. Calling `request.build_absolute_uri` in the Django app gives us the
azurecontainerapps.io URI, whereas we need the public-facing FQDN.

Rather than rely on headers, in this commit we add an explicit BASE_URL
environment variable. This is more visible and easier to debug.

This:

- Should be set to http://locahost:8000 in development
- Should be set to the FQDN in deployed environments

In our system tests, we set BASE_URL to match the value of
live_server_url so that callbacks from our fake CIS2 client work
correctly.