Ensure we don't send notifications on the weekend
Use the default working days from the business days python library to ensure we don't send message batch requests to Notify at the weekend, this reduces the chances of people receiving notifications when the BSOs are closed.
We may want to define a calendar of holidays which is supported by the business days library.