Add opt-in Basic Auth middleware
We need a simple mechanism to restrict access to certain deployed
environments, with the flexibility to mark certain endpoints as free
from this restriction.

This commit:

- Adds a middleware implementation which restricts access
- Makes it opt-in via environment variable
- Adds a decorator to allow certain view functions to skip this check

We mark the jwks_uri endpoint with the decorator, as we want this to be
freely accessible (to the CIS2 auth server, which needs this for
signature verification).