Create seed command for Clinic data
We would like to have the option to preload Clinic
data in order to provide extra location data
(description and google maps link) that is not
available from the data we get in the MESH inbox -
this data may change in the future or we may need
to add Clinics.

I've  created this as needing to be run manually
and with a safety check requiring input, because
this could be changing Production data we should
be extra cautious about when it is run.