add a custom boolean field that renders the checkboxes component
This allows us to have fields that return a boolean, and render as
a single checkbox, equivalent to the built in BooleanField.

An example is when recording a symptom, we have fields like
"The symptom is intermittent" and "The symptom has recently resolved"
which are each individual fields and are not grouped within a fieldset
with a common legend.