refactor presenters into a package
This app is growing big enough that it's useful to start splitting
some of the large modules. I'm aiming for a rails-y structure
where the various structural components are packages with
modules within them.

Have added the suffix _presenter just because it's easier to switch between
files in VS code when the filename is unique.