add clinic location data
the `location_description` and `location_url` will
be used in the personalisation fields when sending
out appointment invites - this JSON data can be
used to cross reference with Clinic data on our db
to get the values. This is a quick solution for
private beta as a way to reference this data
without having to interfere with any existing data
or data sources.