Set CSRF_TRUSTED_ORIGINS
Django's CsrfViewMiddleware validates the Origin header against
the current host and the CSRF_TRUSTED_ORIGINS settings.

It seems like this is failing in the review app environment
because the request header has the frontdoor URL
https://pr-405.manage-breast-screening.non-live.screening.nhs.uk
but the "current host" comes from the container app URL.

We can work around this by explicitly adding the front door URL
as a trusted origin.

https://docs.djangoproject.com/en/5.2/ref/csrf/#how-it-works