Add custom User model
We're about to add an authorization layer based on users, the providers
they're associated with, and the roles they have with those providers.
We're currently using the default Django User model, which isn't intended to
be modified directly and isn't the best fit for how we intend to model
authorization.

There are various suggested approaches for dealing with this, including
simply replacing the default User with a custom one. That's the change
this commit introduces.

Here we:

- Define the new User model.
- Give it the nhs_uid field instead of the username field, as this is
  the unique identifier we'll be using when logging in users via CIS2.
- Continue to support Django Admin (for now at least), by maintaining
  the presence of is_staff and is_superuser fields, using the
  PermissionsMixin, and sustaining the creation of groups/permissions
  tables. We may remove these later to simplify the model
  implementation.
- Update persona login to work with the custom User.
- Add the necessary database migrations.

On migrations - swapping the User model is a fairly invasive change that
conflicts with pre-existing migrated changes. The intention here is to:

- Require a database reset in local dev
- Drop tables in the deployed dev environment and re-run migrations from
  scratch

As we haven't gone live yet, we have no need for a careful migration
approach from the default user table to our custom one.