check `use_fieldset` arg in field templates
Out of the box, django defines per field whether a fieldset should be
used. In our case, we want it to be optional to allow for more
complex form structures where we combine multiple fields into
a single fieldset (so they are displayed under the same legend).

To control this behaviour, introduce some extra widget subclasses that
toggle off the flag.