Set NHS Mail environment variables
These will be set as keyvault secrets on the
Container App Job for CreateReports command.

The SMTP Username and Password will be setup
following the NHS Mail API guidelines:
https://comms-mat.s3.eu-west-1.amazonaws.com/Comms-Archive/applicationsguide.pdf

The password will expire in one year, so we will
need to ensure the keyvault expiry is set on that.

The recipient email will be our private beta BSO
contact in production environment, and a suitable
test email in non-live environments.