add functionality for deleting lumps
after adding a lump, you can click change, and then
there will be a link to delete at the bottom of the form.
This should show a confirmation page first as it's not
easily revertable. On the confirmation page is a button
that POSTs back to the same url. The view will delete
the object and create an audit record.