Add provider selection on log in
- On log in, give the user a choice of which provider to log in with
- Set the chosen provider in a current_provider key in the session
- If only one provider is assigned, skip the selection screen
- If no providers are assigned, show a message informing the user

The 'no providers' screen isn't expected to be seen by production users
given that we expect to pre-populate the database with user accounts and
assignments when rolling out the service. Nonetheless, it's useful to
gracefully handle this scenario rather than error out, particularly in
local dev and on non-live environments.