refactor: extract SymptomForm
There will be forms for every type of symptom and each one will
create/update a Symptom instance.

There are some variations in the fields shown but the behaviour
will be similar across each one. I've moved all of the conditional
validation, and the mapping to/from the model instance to the base
class in the hope that we can just override the field declarations
per symptom type.

If a field isn't on a particular form, it will have no value in
cleaned_data and we will just default that field on the model instance.
When we set the initial values, any that aren't on the form are just
ignored.