use utility classes instead of overriding component styles
The `p:has()` approach is a bit brittle as it will stop working if other
elements come before the inset text.

This is is only used on the record medical info page at the moment, so
we can use utility classes to tweak the margins there. We're already
applying them to the inset text.