Rename fetch_current_provider to fetch_most_recent_provider
Prior to the introduction of the CurrentProviderMiddleware and the
UserAssignment model, the concept of 'current provider' already existed
in the codebase. Here it was relating to a participant's most recent
provider when adding a previous mammogram.

Rename functions and variables related to this concept to help
distinguish it. Leave the ParticipantReportedMammogramForm as is, as it
provides enough context (and is far enough away from the view layer) so
as not to be confusing.