modify system tests to check titles
Where we are otherwise checking that we have landed on a particular
page, ensure that we check the title of the page. This will help
catch cases where we've forgot to set it.

Normally I would check this in the view unit tests, but when using the
Jinja template processor, we are not able to access the context passed
to the view for test client requests.