Remove the superuser concept from the permissions system
The idea of a superuser doesn't really have any purpose in the service
at present, with the currently designed permission model adequately
catered for by the Clinical and Administrative roles that a user can be
assigned.

In the future, a 'superuser' is likely to be a member of our
team with administrative or support access to the service. This can be
handled via the is_superuser flag that currently exists on the User
model.

With all this mind, for now we can remove the concept of a superuser
role and any corresponding access rules.