Rework rules to align with newly designed permission model
Our current permissions model was an early version built around Django
groups.

We've now iterated our requirements for permissions and want to build
out something based on Clinical and Administrative user roles that can
be set on each Provider assignment.

Because we want to set roles per provider, it's easier to build
something custom rather than continue to rely on groups, which don't
align well with the model we have in mind.

Update the rules file and its associated tests to reflect this.