fix: don't allow symptoms to be edited with the wrong form
If there is a bug when editing an existing symptom, the wrong form might
be shown for the symptom type.

Since each change form URL currently works even if the symptom type
doesn't match, such a bug could lead to questions being skipped.

Add type filters to the object lookup so that the form won't be shown in
this case.