Run build step straight away in CI
Our CI currently waits until all tests have run before attempting to
build the image. There is some logic to this - if the tests fail,
you might not want the image to end up in image repository.

However, we would never use one of these images in production, because
the deploy to non-review environments is fully automated, and it's
triggered by a separate workflow that runs against the main branch,
and still requires tests to run first.

I'm removing this requirement so that CI finishes quicker, and we are
not blocked from deploying review apps while the test suite runs.
Sometimes you might want to deploy something that fails some linter or
whatever as this can be done in parallel to people looking at the review
app.