enable docker build cache
Our builds in CI are slow, so we could benefit from caching intermediate
layers.

For a discussion of the different options, see:

- https://docs.docker.com/build/ci/github-actions/cache/
- https://www.blacksmith.sh/blog/cache-is-king-a-guide-for-docker-layer-caching-in-github-actions

mode=max allows intermediate layers to be cached in a multi-stage build,
even if they're not part of the final image.
Since we are using multi-stage builds I think we should have this on.

The GitHub cache option is considering experiemntal, so I've gone with
the registry option.