enable docker build cache
Our builds in CI are slow, so we could benefit from caching intermediate
layers.

mode=max allows intermediate layers to be cached in a multi-stage build,
even if they're not part of the final image[1].
Since we are using multi-stage builds I think we should have this on.

The GitHub cache option is considered experimental, so I've gone with
the registry option, which depends on us using setup-buildx.

As per the build-and-push-docker-images documentation[2], setup-buildx
is generally recommended, and it allows us to use the docker-container
driver[3] instead of the default docker driver when building images.
This driver has the "cache export" functionality we need for registry
caching.

[1]. https://docs.docker.com/build/ci/github-actions/cache/
[2]. https://github.com/marketplace/actions/build-and-push-docker-images
[3]. https://docs.docker.com/build/builders/drivers/