call ApplicationInsightsLogging at run time
We want to ensure that Application Insights
configure method isn't called every time we want
to log to Insights - from the AppConfig Django
docs[1] this looks like the right place to call
setup that we want to happen at the startup of the
app, and it is called before every management
command.

Calling this does mean that high level telemetry
is being reported on the HTTP Requests for the
whole of the Manage app - so e.g. if a request
were made to `/clinics` it would be reported on
Insights, but the generic `stdout` and other logs
won't be. When Manage launches we will want to
call this at the top level config settings to
ensure it's always called, and each team will need
to create their own dashboards to understand
metrics for the different projects and endpoints.

[1] https://docs.djangoproject.com/en/5.2/ref/applications/#django.apps.AppConfig.ready