Remove Django Admin
The admin functionality Django ships with is convenient for changing
data in local development. However:

- Its inclusion in the project adds several tables and fields that we
  don't need, eg - groups, permissions, the is_superuser field on our
  custom User model.
- As we build out the service, the admin needs extending and updating as we
  go.
- It's yet another auth mechanism, alongside Persona and CIS2.
- We don't intend to run it in production

To simplify our codebase, this commit removes:

- Django admin URLs
- Various admin files created for the various Django apps
- The superfluous group and permissions tables for the User app
- The is_superuser field - we can use is_staff instead when checking
  if a user needs elevated permissions

We also delete any records in the auth_groups, auth_group_permissions
and auth_permissions tables. We don't drop these altogether, as they're
tightly coupled with django.contrib.auth.backends.ModelBackend. We keep
the backend as we need to provide a backend identifier when creating a
user session. At some point we may decide to replace this with a
lightweight custom backend, but this isn't needed for now.