fix: include a visually hidden suffix in repeated labels
The "Describe the specific area" label is repeated for each
conditionally shown input.

Colin R pointed out that this can cause an accessibility problem when
using e.g. Dragon Naturally Speaking to control the browser. If multiple
elements on the same page have the same label, it won't know which you
mean. When javascript is working there should only be one of them
visible on the page, but the fields are still there.

Putting the unique part at the end should mean that the user can say the
label as it appears visually, "Describe the specific area" and Dragon
can prompt which one specifically they mean, e.g. "Describe the specific
area: right breast"

This should also workaround our tests in CI for some reason not running
the javascript that hides the non-selected fields. We may need
to look into this further, but it doesn't block this work.